requirements.txt
discord.py
python-dotenv
aiofiles
pyairtable
requests
aiohttp
google-generativeai
chromadb
pytest
